requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
pymongo==4.6.1
python-dotenv==1.0.0
pandas==2.1.4
xlsxwriter==3.1.9
requests==2.31.0
streamlit==1.31.0
plotly==5.18.0
langchain==0.1.0
langchain-openai==0.0.2.post1
openai>=1.6.1,<2.0.0
flask-limiter==3.5.0
flask-cors==4.0.0
redis==5.0.1
Pillow==10.1.0
reportlab==4.0.8
numpy==1.26.2
chromadb==0.4.22
tiktoken==0.5.2
scipy==1.11.4
scikit-learn==1.3.2
flask-jwt-extended==4.5.3
mongoengine==0.27.0
Flask-Babel==2.0.0
Flask-MongoEngine==1.0.0
bcrypt==4.0.1
dnspython==2.4.2
email-validator==2.0.0.post2
gunicorn==20.1.0
python-jose==3.3.0
Flask-Redis==0.4.0